fix(review): Update avg_rating in Course table and fix BindingException

+++ BACK_DangGeunKo/src/main/java/com/danggeunko/review/service/ReviewServiceImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.danggeunko.course.dao.CourseDao;
 import com.danggeunko.review.dao.ReviewDao;
 import com.danggeunko.review.dto.Review;
 import com.danggeunko.review.dto.ReviewSearchCondition;
@@ -13,16 +14,23 @@ import com.danggeunko.review.dto.ReviewSearchCondition;
 public class ReviewServiceImpl implements ReviewService {
 	
 	private final ReviewDao reviewDao;
+	private final CourseDao courseDao;
 	
-	public ReviewServiceImpl(ReviewDao reviewDao) {
+	public ReviewServiceImpl(ReviewDao reviewDao, CourseDao courseDao) {
 		this.reviewDao = reviewDao;
+		this.courseDao = courseDao;
 	}
-
+	private void updateCourseRating(int courseId) {
+        double avgRating = reviewDao.selectAverageRating(courseId);
+        courseDao.updateCourseAvgRating(avgRating, courseId);
+    }
 
 	@Override
 	@Transactional
 	public boolean addReview(Review review) {
-		return reviewDao.insertReview(review) > 0;
+		boolean completed = reviewDao.insertReview(review) > 0;
+		updateCourseRating(review.getCourseId());
+		return completed;
 	}
 
 
@@ -36,14 +44,23 @@ public class ReviewServiceImpl implements ReviewService {
 	@Override
 	@Transactional
 	public boolean updateReview(int reviewId, Review review) {
-		return reviewDao.updateReview(reviewId, review) > 0;
+		boolean completed = reviewDao.updateReview(reviewId, review) > 0;
+		updateCourseRating(review.getCourseId());
+		return completed;
 	}
 
 
 	@Override
 	@Transactional
 	public boolean deleteReivew(int reviewId) {
-		return reviewDao.deleteReview(reviewId) > 0;
+		Review review = reviewDao.getReviewById(reviewId);
+		if (review != null) {
+		boolean completed = reviewDao.deleteReview(reviewId) > 0;
+		updateCourseRating(review.getCourseId());
+		return completed;
+		}
+		return false;
+		
 	}
